refactor: change env to shared in core instead main
