+++ src/main/setup/routes-setup.ts
@@ -3,8 +3,8 @@ import { resolve } from 'node:path'
 
 import { type FastifyInstance } from 'fastify'
 
-import { api } from '@/main/config/env'
-import { getCurrentDirname } from '@/main/config/get-current-dirname'
+import { api } from '@/core/shared/config/env/env'
+import { getCurrentDirname } from '@/core/shared/config/env/get-current-dirname'
 
 export async function routesSetup(app: FastifyInstance): Promise<void> {
   const routesPath = resolve(getCurrentDirname(), '../routes')
